ObjPool 对象池范式

+++ src/main/java/thread/semaphore/pool/ObjPool.java
@@ -40,8 +40,8 @@ class ObjPool<T, R> {
 
     public static void main(String[] args) throws InterruptedException {
         // 创建对象池
-        ObjPool<Long, String> pool =
-                new ObjPool<Long, String>(10, 2L);
+        ObjPool<Integer, String> pool =
+                new ObjPool<Integer, String>(10, 2);
 // 通过对象池获取t，之后执行
         pool.exec(t -> {
             System.out.println(t);
